Extract order by queries tests to separate class
Cherry-pick of https://github.com/prestosql/presto/commit/94b4d0a42f8c86b95b6c3664c5cd83b6eab7297a
